Fourth review pass: six of seven findings, and the one the client disproves
Codex's findings on 9e31f626. Three of them are the same class the FIRST pass claimed
to have fixed -- "exits 0 after doing nothing" -- in stages I fixed the tile instance
of and never went back to. That is the honest summary: the class was named and then
only one member of it was treated.

Applied:

- ExtractDbc answers with the number of databases it could NOT write, the contract
  BakeMap already uses, and the caller aborts before any tile work. Both silent
  continues (mpq.Read failing, fopen failing) and a short fwrite now count, as does
  writing nothing at all -- an empty DBFilesClient glob or a destination that could
  not be created, which produced a clean exit 0 over an empty dbc/.
- A gomodels run that cannot read GameObjectDisplayInfo.dbc logged an error and fell
  through to the success path, leaving gomodels/ empty behind exit 0. Every door,
  lift and bridge in the world is a gomodel.
- The nav .mmap header write returned 0 on failure, and BakeAll treats 0 as a map
  that legitimately produced no tiles -- only negative is fatal. A nav bake could
  finish green with no navmesh for the map. It returns -1 and says why.
- The split-tile loader threw away the _obj0 parse result, so an unreadable object
  ADT gave back correct ground and NO placements, which BakeMap counts as a success
  because it only looks at hasTerrain. That is the exact failure the comment eight
  lines above it warns about.
- The MCLQ bounds check summed uint32, so an offset near UINT32_MAX wraps below span,
  passes the guard, and is then dereferenced. MCVT and MH2O widen to 64 bits; this
  one did not. Test added.
- GameObject::UpdateModel inserted a freshly built model without refreshing its
  collision state, and a fresh model is non-collidable. Only AddToWorld refreshed it,
  so a destructible building swapping to its damaged or destroyed display -- a
  SetDisplayId, not a spawn -- contributed nothing to LOS or height until some later
  state or phase change happened to call it.

NOT applied, because the client disproves it. The finding says a declared WMO group
that ParseWmoGroup rejects must fail the model instead of being skipped. But false
from that function does not mean the group failed to parse: WmoParser.hpp states the
contract in as many words -- "Returns false when the group carries neither collidable
geometry nor liquid" -- and those are the render-only groups the loader deliberately
does not carry, as its own comment says.

I applied it anyway before checking, and measured what it does to real 5.4.8 data:
56 groups across the client answer false, in files of 96KB to 867KB -- Stormwind
group 82 of 281, both Ulduar towers, Icecrown, 24 groups of Skywall -- and NOT ONE
declared group file is missing from the archives. It cost 11 game-object models,
silently downgraded from partial collision to none, and all of map 576 Nexus70,
whose global WMO stopped baking entirely and failed the whole extraction.

The half of the finding that is real is kept: a group file the archive does not have
IS an incomplete client, since the root declares the count. It fails the model now.
The probe shows that case does not occur in retail, so it costs nothing and catches
actual corruption.

Verified on FreeBSD 15 with clang 19.1.7: 117 tests, 0 failed, and a full extraction
of the real 5.4.8 client -- every map in Map.dbc, not a sample -- back to its exact
pre-change baseline: 220 maps, 9,724 tiles, 5,532 gomodels with 2,474 having no
collision, map 576 ok, rc=0.

+++ src/tools/extractor/client/AdtParser.cpp
@@ -182,8 +182,13 @@ namespace world::terrain
             // MCLQ is the pre-WotLK liquid chunk and is absent from 3.3.5a client data;
             // it is read only so an older or hand-made tile is not silently dry. MH2O,
             // parsed after every MCNK, overrides whatever this writes.
-            constexpr uint32_t MCLQ_BYTES = 8 + 81 * 8 + 64;
-            if (!offsMclq || sizeMclq <= 8 || offsMclq + 8 + MCLQ_BYTES > span)
+            // Summed in 64 bits: offsMclq comes straight off the file, and a value near
+            // UINT32_MAX makes the 32-bit sum WRAP BELOW span, so the guard passes and
+            // mcnk + offsMclq reads outside the buffer. MCVT and MH2O widen for the same
+            // reason; this one did not.
+            constexpr uint64_t MCLQ_BYTES = 8 + 81 * 8 + 64;
+            if (!offsMclq || sizeMclq <= 8 ||
+                uint64_t(offsMclq) + 8 + MCLQ_BYTES > uint64_t(span))
             {
                 return true;      // no MCLQ is the 5.4.8 norm, and the heights are read
             }
